refactor: move json path stripping to output crate
